Move strorage config into config pkg
